Get rid of unnecessary `ConvertOrderStatusToRejectOrderCode`
